refactor(arm64): isolate NEON compare selection
